Week 2 Que - Find duplicates, Shuffle array
Session - 2 | 24 Sept 2022 | IST

Q1: Array Shuffling: Given an array of integers, Shuffle the numbers in such a way that all even numbers should be in the first half of the array, and all odd numbers should be in the 2nd half of the array.
For example, if the Input array is { 3, 5, 9, 6, 2, 8, 11}, the algorithm can return either {6, 2, 8, 3, 5, 9, 11} or {8, 2, 6, 9, 5, 3, 11}. Both are valid answers.

Q2: Find any duplicate number: An array contains n numbers ranging from 0 to n-1. There are some numbers duplicated in the array. It is not clear how many numbers are duplicated or how many times a number gets duplicated. How do you find a duplicated number in the array?
For example, if an array of length 7 (n=7) contains the numbers {2, 3, 1, 0, 2, 5, 3}, the implemented function (or method) should return either 2 or 3.