Week 3 Questions
Q1: Merge Sorted arrays using a separate array: Given two Sorted arrays, create a third array from the elements of both the arrays in sorted order. -
Q2: Merge Sorted arrays in place: Given two Sorted arrays( A and B). Array A has unlimited memory or enough memory to hold the elements of B. The current number of elements in A was also given as input. Sort the elements of A and B without using the 3rd array.